source_bag/requirements.txt
dbt-core
dbt-snowflake
dbt-databend-cloud
snowflake-connector-python
pymysql
python-dotenv
